requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
httpx
yfinance
openai
google-genai
anthropic
apscheduler
